major-change: better functionality and added compatibility with mp3

+++ audio_data.py
@@ -1,10 +1,11 @@
 """
-This will read the m4a files in the cwd into a list, and open each file with mutagen
+This will read all files in the cwd into a list, and open each file with mutagen if mp3, mp4 or m4a
 Split the filenames into artist and title and add them to the metadata
 """
 import re
 import os
 from mutagen.easymp4 import EasyMP4
+from mutagen.easyid3 import EasyID3
 
 
 class AudioFiles:
@@ -14,39 +15,65 @@ class AudioFiles:
         self.music_files = []
         self.music_artist = ""
         self.music_title = ""
+        self.artist_title = ""
         self.tags = []
 
     def set_metadata(self):
         for i, f in enumerate(os.listdir(self.dir_w_files)):
-            self.music_files.append(f)
+            if os.path.isfile(f):
+                self.music_files.append(f)
 
         for i in self.music_files:
-            audiofile = EasyMP4(i)
-            artist_title = i.split(" - ", 1)
-            self.music_artist = artist_title[0]
-            self.music_title = artist_title[1][:len(artist_title[1])-15]
+            if ".m4a" in i or ".mp4" in i:
+                audiofile4 = EasyMP4(i)
+                self.artist_title = i.split(" - ", 2)
+                self.music_artist = self.artist_title[0]
+                self.music_title = self.artist_title[1][:len(self.artist_title[1]) - 4]
 
-            audiofile["artist"] = f"{self.music_artist}"
-            audiofile["title"] = f"{self.music_title}"
-            audiofile.save()
+                audiofile4["artist"] = f"{self.music_artist}"
+                audiofile4["title"] = f"{self.music_title}"
+                audiofile4.save()
+
+            if ".id3" in i or ".mp3" in i:
+                try:
+                    audiofile3 = EasyID3(i)
+                    self.artist_title = i.split(" - ", 2)
+                    self.music_artist = self.artist_title[0]
+                    self.music_title = self.artist_title[1][:len(self.artist_title[1]) - 4]
+
+                    audiofile3["artist"] = f"{self.music_artist}"
+                    audiofile3["title"] = f"{self.music_title}"
+                    audiofile3.save()
+                except ValueError:
+                    pass
 
     def get_metadata(self):
         for i, f in enumerate(os.listdir(self.dir_w_files)):
-            self.music_files.append(f)
+            if os.path.isfile(f):
+                self.music_files.append(f)
 
         for i in self.music_files:
-            audiofile = EasyMP4(i)
-            self.tags.append(audiofile)
+            if ".id3" in i or ".mp3" in i:
+                try:
+                    audiofile3 = EasyID3(i)
+                    self.tags.append(audiofile3)
+                except ValueError:
+                    pass
+            if ".m4a" in i or ".mp4" in i:
+                audiofile4 = EasyMP4(i)
+                self.tags.append(audiofile4)
         return self.tags
 
     def prepend(self, prep):
         for i, f in enumerate(os.listdir(self.dir_w_files)):
-            src = os.path.join(self.dir_w_files, f)
-            dst = os.path.join(self.dir_w_files, f"{prep}{f}")
-            os.rename(src, dst)
+            if os.path.isfile(f):
+                src = os.path.join(self.dir_w_files, f)
+                dst = os.path.join(self.dir_w_files, f"{prep}{f}")
+                os.rename(src, dst)
 
     def remove_substr(self, substr):
         for i, f in enumerate(os.listdir(self.dir_w_files)):
-            src = os.path.join(self.dir_w_files, f)
-            dst = os.path.join(self.dir_w_files, re.sub(f"{substr}.*?", "", f))
-            os.rename(src, dst)
+            if os.path.isfile(f):
+                src = os.path.join(self.dir_w_files, f)
+                dst = os.path.join(self.dir_w_files, re.sub(f"{substr}.*?", "", f))
+                os.rename(src, dst)
